requirements.txt
# Library dependencies for the python code.  You need to install these with
# `pip3 install -r requirements.txt` before you can run this.

pymongo==3.10.0
mongomock==3.18.0
PyQt5==5.14.0
PyQt5-sip==12.7.0
helpdev==0.6.10
importlib-metadata==1.3.0
more-itertools==8.0.2
psutil==5.6.7
qdarkgraystyle==1.0.2
zipp==0.6.0
python-dateutil==2.8.1
coverage==5.0.3